requirements.txt
PyYAML>=6.0.1
pandas>=2.2.2
matplotlib>=3.9.0

